Update useSharedValue.mdx
Updates useSharedValue page to include an example where a useSharedValue is passed down as props.

If there is an issue with what I have proposed, please let me know if I can change it and resubmit. Thank you all so much!!!

Further aside as to why I think this change is necessary:

After reading this page plenty of times over the last year, I had no idea about this parent-child component behavior specifically. The way the docs currently read, it almost sounds like the opposite, with a quote like: "won't trigger a typical React re-render." Even though I know this is true for the JS/React relationship, without further explanation it leads me the believe that if I pass a shared value down as props, it won't update for the child JSX. I thought I always would have to confine useSharedValue to a single component and can't ever pass it down to child components. I would have found an example like this very helpful and others may as well.